ApplicationServer/ml_server/requirements.txt
fastapi>=0.114
uvicorn[standard]>=0.30
transformers>=4.44
torch>=2.3
pillow>=10.3
timm>=1.0.7
accelerate>=0.33
python-multipart>=0.0.9